First function complete
The search form is now able to get company information. A loop is first used in case there is more than one company matching the first 3 charactors of a company number, for example. An if statement is then used in order to capture the status of the company which will help with scoring later on.